lvlup-api/requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.2
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
Pillow==7.1.1
psycopg2==2.8.5
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
wrapt==1.11.2
cryptography~=2.8
drf-jwt~=1.13.3
pyjwt~=1.7.1
requests~=2.22.0